pyproject.toml
[tool.poetry]
name = "pubmed-paper-fetcher"
version = "1.0.0"
description = "Fetch PubMed papers with company authors"
authors = ["Gour Krishna Dey <your@email.com>"]
readme = "README.md"
license = "N/A"
packages = [
    { include = "pubmed_fetcher", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "*"
biopython = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"

[tool.poetry.scripts]
get-papers-list = "pubmed_fetcher.cli:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
